Updated workout routes and curl-scripts
Made changes to the SHOW route for showing one workout, so that the workout can be found by name, and without a token. Updated INDEX route to show all WODs without a token.

Soon to add relationships for user comments on workouts to add personal scores and stats.

Still need to make update and delete WOD work.